added auth to all the endpoints of the user, mde it so that users can update the journal entries pr delete the journal entries by the user ID

+++ src/main/java/com/example/crud/Service/JournalEntryService.java
@@ -49,10 +49,21 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id); // returns the founded ID if found else returns null, so optional works here
     }
 
-    public void deleteById(ObjectId id, String userName) {
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveNewUser(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if(removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+                }
+            }
+            catch (Exception e) {
+                System.out.println(e);
+                throw new RuntimeException("An error has been occurred while deleting journal entry");
+        }
+        return removed;
     }
 }
